refactor(test): meet the FR-006 line-cap ratchet via shared helpers
The round-2 fix commit grew six over-cap test files, tripping the
check:test-file-cap growth ratchet in CI. Extract the shared behavior:

- tests/helpers/wait-for.ts: the attempt-counting polling wait (with its
  unit suite moved to tests/helpers/wait-for.test.ts) out of the
  Linux-gated plan-sync suite.
- tests/helpers/discovery-seam.ts: installComposerProbeSeam/
  restoreDiscoverySeam shared by both pkg-audit-composer suites (and
  formerly-duplicated seam plumbing), removing vestigial spawnCalls
  tracking from the adversarial suite.
- invariant-gate-fixtures: buildInvariantsOracleExpected() dedupes the
  legacy-oracle's two expected-output arrays (six-check block shared).
- comment/blank-line trim in knowledge-query and both phase-complete
  suites. All six ratcheted files are back at or below their baseline.
